# bpo-33387: Simplify bytecodes for try-finally, try-except and with blocks.