… Lib/test/test_dis.py
… Objects/frameobject.c
… Python/compile.c
… Python/importlib.h
… Python/importlib_external.h
… Python/importlib_zipimport.h
… Python/opcode_targets.h