Quoting the bpo issue:
```
The six complex bytecodes currently used for implementing 'with' and 'try' 
statements can be replaced with just two simpler bytecodes.
The six bytecodes are WITH_CLEANUP_START, WITH_CLEANUP_FINISH,
BEGIN_FINALLY, END_FINALLY, CALL_FINALLY and POP_FINALLY.
They can be replaced with RERAISE and WITH_EXCEPT_START.

See https://bugs.python.org/issue32949 for more details of the new bytecodes and 
how they are used in the 'with' statement.

The try-finally statement can be implemented broadly as 
   SETUP_FINALLY except
   try-body
   POP_BLOCK
   finally-body
   JUMP exit
except:
   finally-body
exit:
```

The implementation of `with` statements is a bit more complex. The with statement

```
   with EXPR as VAR:
       BLOCK
```
   is implemented as:
```
        <code for EXPR>
        SETUP_WITH  ex
        <code to store to VAR> or POP_TOP
        <code for BLOCK>
        LOAD_CONST (None, None, None)
        CALL_FUNCTION_EX 0
        JUMP_FORWARD  done
    ex:  WITH_EXCEPT_START (calls EXPR.__exit__)
        POP_JUMP_IF_TRUE t:
        RERAISE
    t:  POP_TOP * 3 (remove exception from stack)
        POP_EXCEPT
        POP_TOP
    done:
```




<!-- issue-number: bpo-33387 -->
https://bugs.python.org/issue33387
<!-- /issue-number -->