@methane 2018-07-11
@markshannon FYI, `perf compare_to old.json new.json -G` (-G option) makes output more compact and readable.