@markshannon 2018-09-19
All comments addressed and rebased to fix merge conflict with https://github.com/python/cpython/commit/ddd1949fea59f256e51191540a4446f75ed608fa