@serhiy-storchaka 2019-05-19
Only the qualified review is needed. The problem is that the code itself is very complex, and we have not much core developers well known with this code besides @markshannon and me. I would prefer to make changes in few simpler steps. It is not so easy in this case, because several parts are mutually related.