@markshannon 2019-05-26
Needed rebasing yet again, due to unrelated changes in Python/ceval.c